Moved some common elements into the common components folder and out of the pokedex one
