+++ src/app/teambuilder/page.tsx
@@ -11,6 +11,7 @@ import type { NextPage } from "next";
 import Head from "next/head";
 import { useEffect, useMemo, useState } from "react";
 import PokemonCard from "../../components/PokemonCard";
+import TypeChartCell from "../../components/TypeChartCell";
 import { decodeTeam, encodeTeam, MAX_LEVEL, SavedPartyPokemon } from "../data/teamExport";
 import { Ability } from "../data/tectonic/Ability";
 import { Item } from "../data/tectonic/Item";
@@ -21,7 +22,6 @@ import { TectonicData } from "../data/tectonic/TectonicData";
 import { calcBestMoveMatchup, calcTypeMatchup } from "../data/typeChart";
 import { PartyPokemon } from "../data/types/PartyPokemon";
 import { isNull } from "../data/util";
-import TypeChartCell from "../pokedex/components/TypeChartCell";
 import AtkTotalCell from "./components/AtkTotalCell";
 import DefTotalCell from "./components/DefTotalCell";
 import MatchupMonCell from "./components/MatchupMonCell";
