+++ src/components/EncounterDisplay.tsx
