+++ src/components/StatRow.tsx
