+++ src/components/TypeChartCell.tsx
